feat(pull): allow pulling before pushing
When running several tasks in parallel in forks of the same remote, it is normal that one of them fails.

With this new feature, you can instruct the module to pull upstream changes before pushing. It's also easy to use Ansible's loop-until feature to make several attempts until one succeeds.

@moduon MT-1075